Over about a 10s period in o!f (`TestSceneVertexUploadPerformance`):

| before | after |
| --- | --- |
| <img width="833" alt="image" src="https://github.com/user-attachments/assets/5c49786c-784f-41f2-ad24-2b4796da8cc5"> | <img width="832" alt="image" src="https://github.com/user-attachments/assets/e4e56563-fc38-4fbe-9d4f-712b1778ce21"> |